Game-Whoever-says-20-is-a-loser
It is a game of whoever says twenty loses; the user challenges the computer, as the possibility of the computer losing is almost impossible thanks to an algorithm that it discovered that avoids numbers that lead to the player losing. I leave you with the program to experience the competition with the computer.